Update Server-Setup-and-Maintenance-Linux.md
Replaced topaz with LSB or xi (where relevant for either) in the Update section and the Rasperry Pi Warnings section, not sure why I didn't catch them in my previous pass.

I'm not finding any other mentions of Topaz/Darkstar/TPZ/DSP in any other files (aside from mentions that are meant to be there) so this should be the last PR I do for this.  Topaz can now be put out to pasture.